Separate domain and infrastructure namespaces
