+++ Parkman/Domain/Entities/CompanyProfile.cs
@@ -1,6 +1,6 @@
 using System;
 
-namespace Parkman.Domain;
+namespace Parkman.Domain.Entities;
 
 public class CompanyProfile
 {
