+++ Parkman/Domain/Entities/Vehicle.cs
@@ -1,6 +1,8 @@
 using System;
 
-namespace Parkman.Domain;
+using Parkman.Domain.Enums;
+
+namespace Parkman.Domain.Entities;
 
 public class Vehicle
 {
